Add Node#to_paths
